Rearrange .sqlalchemy_utils.fields content into separated files
