--- alfred/sqlalchemy_utils/fields/__init__.py
@@ -1,8 +1,0 @@
-from .image import ImageType
-from .password import PasswordSaltType, PasswordType
-
-__all__ = [
-    "ImageType",
-    "PasswordSaltType",
-    "PasswordType",
-]
